requirements.txt
fastapi
uvicorn[standard]
databases[postgresql]
asyncpg
pydantic
pytest
httpx
python-dotenv
pytest-asyncio